+++ test/transformers/test_embedding.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 from torch.nn import Embedding
 
-from liger_kernel.transformers.embedding import LigerEmbedding
+from liger_kernel.transformers.experimental.embedding import LigerEmbedding
 
 SLEEP_SECONDS = 0.1
